wallettemplate/build.gradle: add mergedModule requires
The mergedModule is using and _should_ `require`:

* `java.logging`
* `jdk.unsupported`
* `org.slf4j`

but wallettemplate has worked fine without it until we run with
Temurin JDK 24 or 25 (which requires the BA Jlink Plugin 3.1.4 to build)
and then there are fatal exceptions at runtime when these modules
are accessed.

This fix should work with all currently supported JDK versions and
distributions.