wallettemplate/build.gradle: add mergedModule block
This commit provides a mergedModule block (inside the `jlink` block)
for the badass-jlink-plugin to define the combined dependencies
of the 4 non-modular JARs currently included in wallettemplate:

* bitcoinj-base
* bitcoinj-core
* Bouncy Castle
* ProtoBuf

The mergedModule `require`s:

* `java.logging`
* `jdk.unsupported`  (used by ProtoBuf)
* `org.slf4j`

with Temurin 24+ (specifically JEP 493) it is now necessary to
provide the mergedModule definition rather than rely on the
automatically-generated one.

For more information about the mergedModule block see the
badass-jlink-plugin documentation.