gradle.yml: install JDK 8 via setup-java action
We previously relied on it being already installed on the GitHub runner, but
we need to upgrade the macOS runner to macOS-15-intel which does not have
JDK 8 built-in.

Note that since we have configured setup-java to use the Temurin distro,
JDK 8 will now be the Temurin build rather than whatever was included
with the runner.

We use an environment variable `MIN_JDK` to specify that these minJDK tests
are to use Java 8, but using a variable allows us to bump the version or
even make it conditional (like we will need to do to add macOS aarch64
runners).