LegacyAddress, SegwitAddress: make field `networks` really immutable
`enum` types are supposed to be immutable, but `EnumSet` isn't.

In each case, it's in a private field and is not mutated, but just to be
a little more clear and defensive, let's wrap the EnumSets in
`Collections.unmodifiableSet()`. This also changes the type of the fields
to the `Set` interface.