LegacyAddress,SegwitAddress: use unmodifiableSet
`enum`s are supposed to be immutable, but EnumSet is mutable.

In each case, it's in a private field and is not mutated, but just to be
a little more clear and defensive, let's wrap the EnumSets in `unmodifiableSet`.
This also changes the type of the private fields to the `Set` interface.